refactor(RoomView): move anchor resolver into services and flag equal-ts jump collisions
The pure anchor resolver lived under List/hooks but is imported by the services/ jump orchestrator, inverting the dependency. Move it (and its test) into services/ so the layering matches the import direction.

Narrow anchorForServerChunk's ts param to Date | number and normalize the string form once at the services edge.

Under __DEV__, warn when the resolved scalar-ts window bound is shared by 2+ rows: the bound can't split equal-ts rows, so the jump may land on the wrong message. The full fix is composite (ts,id) ordering.
